Move app-monitoring-data-ingest to root
