--- app-monitoring-data-ingest/.env.example
@@ -1,7 +1,0 @@
-PG_PORT=5432
-PG_USERNAME=
-PG_PASSWORD=
-PG_HOST=localhost
-PG_DATABASE=
-MONITORING_USERNAME=
-MONITORING_PASSWORD=
