--- app-monitoring-data-ingest/.gitignore
@@ -1,5 +1,0 @@
-.vscode/
-node_modules/
-secrets/
-*.sig
-.env
