--- app-monitoring-data-ingest/Dockerfile
@@ -1,5 +1,0 @@
-FROM node:16
-WORKDIR /usr/app
-COPY . .
-RUN npm ci
-CMD ["node", "index.js"]
