--- app-monitoring-data-ingest/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "app-monitoring-data-ingest",
-  "version": "1.0.0",
-  "description": "Scrapes data from CHT instances and ingests into Postgres",
-  "main": "index.js",
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "btoa": "^1.2.1",
-    "dotenv": "^16.0.0",
-    "node-fetch": "^2.6.0",
-    "lodash": "^4.17.15",
-    "pg": "^8.7.1"
-  },
-  "devDependencies": {
-    "chai": "^4.2.0",
-    "chai-exclude": "^2.0.2",
-    "mocha": "^7.1.2"
-  }
-}
