+++ postgres/functions/create_doc_error_index.sql
