+++ postgres/functions/create_doc_type_task_index.sql
