+++ postgres/functions/get_doc_error_aggr.sql
