+++ postgres/functions/get_forms_by_completion.sql
