+++ postgres/functions/get_invalid_doc_count.sql
