+++ postgres/functions/get_task_by_state.sql
